Fuck it, we'll do it live
Created an entirely new menu hierarchy again. This time, using multiple
scenes instead of panels. Still uses NGUI for sprites and labels and
things, but doesn't use the event system.

I decided to make my own event system. With blackjack, and hookers.

But really though, wrote a subclass of the Input Manager to dole out
commands to the menus. Finally managed to grok messages and object
hierarchies enough to make the menu code not a complete clusterfuck.

Settings and Module menus to come.